use mx3 logger
