+++ mx3_beamline_library/config.py
@@ -1,32 +1,10 @@
-import logging
 from os import environ, path
 
 from redis import StrictRedis
 from redis.exceptions import ConnectionError
 from yaml import safe_load
 
-
-# Logger config
-def setup_logger():
-    # Create a logger
-    logger = logging.getLogger(__name__)
-    logger.setLevel(logging.DEBUG)
-
-    # Create a console handler and set the level to info
-    console_handler = logging.StreamHandler()
-    console_handler.setLevel(logging.INFO)
-
-    # Create a formatter and attach it to the console handler
-    formatter = logging.Formatter(
-        "%(asctime)s %(levelname)s: %(message)s", datefmt="%d-%m-%Y %H:%M:%S"
-    )
-    console_handler.setFormatter(formatter)
-
-    # Add the console handler to the logger
-    logger.addHandler(console_handler)
-
-    return logger
-
+from .logger import setup_logger
 
 logger = setup_logger()
